Fix Abiquo updates repository issues
Enable Abiquo updates repository if it is available for the current version;
otherwise the playbook finishes with error because it can not install any
package (that is, it aborts on the first call to 'yum').

This is the same commit as the commit e5b2b8b08416926a71643c88a2dfe1f86da184cb
from the 'etisalat-uat-environment' repository. For more information see:
    https://github.com/abiquo/etisalat-uat-environment/commit/e5b2b8b08416926a71643c88a2dfe1f86da184cb